all_requirement_neo.txt
absl-py==0.5.0
apixu-client==0.0.1
APScheduler==3.3.1
astor==0.7.1
attrs==18.2.0
Automat==0.7.0
boto3==1.4.4
botocore==1.5.95
certifi==2018.8.24
chardet==3.0.4
click==6.7
cloudpickle==0.5.6
colorama==0.3.9
colorclass==2.2.0
coloredlogs==10.0
colorhash==1.0.2
ConfigArgParse==0.13.0
constantly==15.1.0
cycler==0.10.0
cymem==1.31.2
cytoolz==0.9.0.1
decorator==4.3.0
dill==0.2.8.2
docopt==0.6.2
docutils==0.14
en-core-web-sm==1.2.0
fakeredis==0.10.3
fbmessenger==5.3.0
Flask==1.0.2
Flask-Cors==3.0.6
Flask-JWT-Simple==0.0.3
ftfy==4.4.3
future==0.16.0
gast==0.2.0
gevent==1.2.1
graphviz==0.9
greenlet==0.4.15
grpcio==1.15.0
h5py==2.7.0
html5lib==1.0.1
humanfriendly==4.16.1
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
jsonpickle==0.9.4
jsonschema==2.6.0
Keras==2.0.8
kiwisolver==1.0.1
klein==17.2.0
languageflow==1.1.7a3
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.3
mattermostwrapper==2.1
msgpack==0.5.6
msgpack-numpy==0.4.3.2
murmurhash==0.28.0
networkx==2.2
numpy==1.15.3
packaging==17.1
pandas==0.23.4
pandoc==1.0.0b2
pathlib==1.0.1
patsy==0.5.0
pika==0.11.2
plac==0.9.6
ply==3.11
preshed==1.0.1
prompt-toolkit==1.0.14
protobuf==3.6.1
Pygments==2.2.0
PyHamcrest==1.9.0
PyInquirer==1.0.2
PyJWT==1.6.4
pykwalify==1.6.0
pymongo==3.7.1
pypandoc==1.4
pyparsing==2.2.1
pyreadline==2.1
PySocks==1.6.8
python-crfsuite==0.9.5
python-dateutil==2.7.3
python-engineio==2.2.0
python-socketio==2.0.0
python-telegram-bot==10.1.0
pytz==2018.5
pyvi==0.0.9.1
PyYAML==3.13
rasa-core==0.11.12
rasa-core-sdk==0.11.4
rasa-nlu==0.13.7
redis==2.10.5
regex==2017.4.5
requests==2.19.1
rocketchat-API==0.6.22
ruamel.yaml==0.15.34
s3transfer==0.1.13
scikit-learn==0.19.1
scipy==1.1.0
simplejson==3.11.1
six==1.10.0
sklearn-crfsuite==0.3.5
slackclient==1.3.0
spacy==2.0.12
statsmodels==0.9.0
tabulate==0.8.2
tensorboard==1.10.0
tensorflow==1.10.0
termcolor==1.1.0
terminaltables==3.1.0
thinc==6.10.3
toolz==0.9.0
tqdm==4.11.2
twilio==6.17.0
Twisted==18.7.0
typing==3.5.3.0
tzlocal==1.5.1
ujson==1.35
underthesea==1.1.8
urllib3==1.23
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.53.0
Werkzeug==0.14.1
wrapt==1.10.11
zope.interface==4.5.0
